fix: X11 problems resulting in STATEBYLINE not generated
If a user starts up a SLURM job on a remote server that has X11
forwarding and the user disconnects from said server, ChromHMM will
throw an error:

Exception in thread "main" java.awt.AWTError: Can't connect to X11 window
server using 'localhost:10.0' as the value of the DISPLAY variable.

This will unset $DISPLAY, meaning this error will not appear and
STATEBYLINE can still be generated.